style: OAuth 등록/탈퇴 클래스명 변경
