+++ src/main/java/coffeemeet/server/oauth/implement/client/OAuthMemberClientRegistry.java
@@ -1,8 +1,9 @@
-package coffeemeet.server.oauth.infrastructure;
+package coffeemeet.server.oauth.implement.client;
 
 import static coffeemeet.server.auth.exception.AuthErrorCode.INVALID_LOGIN_TYPE;
 
 import coffeemeet.server.common.execption.InvalidAuthException;
+import coffeemeet.server.oauth.domain.OAuthMemberDetail;
 import coffeemeet.server.user.domain.OAuthProvider;
 import java.util.Map;
 import java.util.Optional;
@@ -12,23 +13,23 @@ import java.util.stream.Collectors;
 import org.springframework.stereotype.Component;
 
 @Component
-public class OAuthUnlinkClientComposite {
+public class OAuthMemberClientRegistry {
 
   private static final String INVALID_LOGIN_TYPE_MESSAGE = "로그인 타입(%s)에 일치하는 타입이 없습니다.";
 
-  private final Map<OAuthProvider, OAuthUnlinkClient> mapping;
+  private final Map<OAuthProvider, OAuthMemberClient> mapping;
 
-  public OAuthUnlinkClientComposite(Set<OAuthUnlinkClient> details) {
-    this.mapping = details.stream().collect(
-        Collectors.toUnmodifiableMap(OAuthUnlinkClient::oAuthProvider, Function.identity())
+  public OAuthMemberClientRegistry(Set<OAuthMemberClient> clients) {
+    this.mapping = clients.stream().collect(
+        Collectors.toUnmodifiableMap(OAuthMemberClient::oAuthProvider, Function.identity())
     );
   }
 
-  public OAuthUnlinkDetail unlink(OAuthProvider oAuthProvider, String accessToken) {
-    return getClient(oAuthProvider).unlink(accessToken);
+  public OAuthMemberDetail fetch(OAuthProvider oAuthProvider, String authCode) {
+    return getClient(oAuthProvider).fetch(authCode);
   }
 
-  private OAuthUnlinkClient getClient(OAuthProvider oAuthProvider) {
+  private OAuthMemberClient getClient(OAuthProvider oAuthProvider) {
     return Optional.ofNullable(mapping.get(oAuthProvider))
         .orElseThrow(() -> new InvalidAuthException(
             INVALID_LOGIN_TYPE,
