+++ src/main/java/coffeemeet/server/oauth/implement/client/OAuthUnlinkClient.java
@@ -1,5 +1,6 @@
-package coffeemeet.server.oauth.infrastructure;
+package coffeemeet.server.oauth.implement.client;
 
+import coffeemeet.server.oauth.infrastructure.OAuthUnlinkDetail;
 import coffeemeet.server.user.domain.OAuthProvider;
 
 public interface OAuthUnlinkClient {
